test+fix: 28 unit tests + 0 warnings (28→0) (#15)
* test: thêm DhcbTools.Tests — 28 unit tests logic thuần C# (không cần Revit)

* fix: 0 warnings — nullable fixes, ElementId(long), PromptForFolder nullable return

+++ src/DhcbTools.Core/ParameterSync/ParameterImportCommand.cs
@@ -55,7 +55,7 @@ public sealed class ParameterImportCommand : ICoreCommand<ParameterImportConfig>
 #if NET8_0_WINDOWS
             var elementId = new ElementId(idValue);
 #else
-            var elementId = new ElementId((int)idValue);
+            var elementId = new ElementId((long)idValue);
 #endif
             var element = document.GetElement(elementId);
             if (element is null)
